pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "astrogeo-lite-auto"
version = "0.1.0"
description = "Automatic astronomy-and-geometry MVP for estimating photo latitude/longitude from stars and building verticals."
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { file = "LICENSE" }
authors = [{ name = "AstroGeo Lite Auto contributors" }]
dependencies = [
  "numpy>=1.24,<2.1",
  "scipy>=1.10,<1.14",
  "opencv-python>=4.8,<5",
  "Pillow>=10,<12",
  "exifread>=3.0",
  "astropy>=5.3,<7",
  "matplotlib>=3.7,<3.10",
  "typer>=0.9,<1",
  "pydantic>=2,<3",
  "requests>=2.31",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
ui = ["streamlit>=1.30"]

[dependency-groups]
dev = ["pytest>=8.0"]

[project.scripts]
astrogeo = "astrogeo.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"
